+++ gui/src/components/Layout.tsx
@@ -213,7 +213,10 @@ const Layout = () => {
   );
 
   useEffect(() => {
-    if (isNewUserOnboarding() && location.pathname === "/") {
+    if (
+      isNewUserOnboarding() &&
+      (location.pathname === "/" || location.pathname === "/index.html")
+    ) {
       dispatch(setOnboardingCard({ show: true, activeTab: "Quickstart" }));
     }
   }, [location]);
