+++ gui/src/components/OnboardingCard/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./useCheckOllamaModels";
+export * from "./useCompleteOnboarding";
